+++ client/src/pages/Dashboard.js
@@ -1,9 +1,14 @@
 import React from 'react';
+import '../styles/Dashboard.css';
 
+/**
+ * This page acts as a dashboard for the app. Will showcase multiple pieces of different pages and act as a functional sneak-peek.
+ */
 const Dashboard = () => {
     return (
         <div className="dashboard">
             <h1>Dashboard</h1>
+            <h2>WIP</h2>
         </div>
     );
 };